# Documentation/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillShare Network",
    "description": "REST API платформы SkillShare Network. Большинство эндпоинтов требуют заголовок `Authorization: Bearer <JWT>`. Получите токен через `POST /api/v1/auth/login`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Вход в систему",
        "description": "Проверяет email и пароль, возвращает JWT access_token для заголовка Authorization.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос (400 Bad Request)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Неверный текущий пароль"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации (422 Unprocessable Entity)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Пароль должен содержать минимум 10 символов"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован (401 Unauthorized)",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing or invalid Authorization header"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT access token",
            "examples": [
              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            ]
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Тип токена (OAuth2 Bearer)",
            "default": "bearer",
            "examples": [
              "bearer"
            ]
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "description": "JWT-токен для заголовка Authorization: Bearer ..."
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Email",
            "examples": [
              "user@example.com"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Пароль",
            "examples": [
              "SecurePass123"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "description": "Тело запроса входа."
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT из ответа POST /api/v1/auth/login"
      }
    }
  },
...
}